feat: voteCounts 쿼리 수정

+++ src/main/java/com/ODG/ODG_back/dto/vote/response/VoteResultDto.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class VoteResultDto {
     private Long placeId;
     private String name;
-    private int voteCount;
+    private Long voteCount;
 }
